Make the pipeline work better

+++ include/exec_user_hooks.py
@@ -2,21 +2,22 @@ import os
 from .exec_in_chroot import exec_in_chroot
 from .gentoomuch_common import user_scripts_path, user_removes_path, chroot_workdir
 from .sign_stage import sign_stage
+from .verify_tarball import verify_tarball
 
 def exec_user_hooks(stage_path: str, scripts: [str], removes: [str]):
+    if not verify_tarball(stage_path):
+        print("EXEC USER HOOKS: COULD NOT VERIFY STAGE")
+        return False
     code = os.system('mkdir -p ' + chroot_workdir + ' && rm -rf ' + os.path.join(chroot_workdir, '*') + ' && tar xpf ' + stage_path + ' -C ' + chroot_workdir)
     if code == 0:
         pass
     if len(scripts) > 0:
         scripts_fullpath = []
-        i = 0
-        while i < len(scripts) - 1:
-            if scripts[i] == "README.md":
+        for s in scripts:
+            if s == "README.md":
                 continue
-            scripts_fullpath.append(os.path.join('/mnt/user.scripts', scripts[i]) + ' && ')
-            i += 1
-        scripts_fullpath.append(os.path.join('/mnt/user.scripts', scripts[-1])) # This time without the trailing && 
-        results = exec_in_chroot(chroot_workdir, ''.join(scripts_fullpath))
+            scripts_fullpath.append(os.path.join('/mnt/user.scripts', s))
+        results = exec_in_chroot(chroot_workdir, ' && '.join(scripts_fullpath))
         if not results:
             print("EXEC USER HOOKS: Chroot command failed")
             # code = os.system('sudo rm -rf ' + chroot_workdir)
@@ -35,12 +36,12 @@ def exec_user_hooks(stage_path: str, scripts: [str], removes: [str]):
             for ll in list_contents:
                 removes_fullpath.append(os.path.join(chroot_workdir, ll))
         remove_commands = []
-        i = 0
-        while i < len(removes_fullpath) -1:
-            remove_commands.append('rm -rf ' + removes_fullpath[i] + ' && ')
-            i += 1
-        remove_commands.append('rm -rf ' + removes_fullpath[-1])
-        code = os.system(''.join(remove_commands))
+        for r in removes_fullpath:
+            if r == "README.md":
+                continue
+            remove_commands.append('rm -rf ' + r)
+        print("EXEC USER HOOKS: Removing " + " ".join(removes_fullpath))
+        code = os.system(' && '.join(remove_commands))
         if not code == 0:
             print("EXEC USER HOOKS: Removes failed")
             # code = os.system('sudo rm -rf ' + chroot_workdir)
